refactor/be/#157 Admin모듈 Manager 기능 분리
