+++ api/src/main/java/kernel/maidlab/api/manager/repository/ManagerRepository.java
@@ -14,5 +14,4 @@ public interface ManagerRepository extends JpaRepository<Manager, Long>, Manager
 
 	Optional<Manager> findByPhoneNumber(String phoneNumber);
 
-	Page<Manager> findAllByIsVerified(Status status, Pageable pageable);
 }
