+++ api/src/main/java/kernel/maidlab/api/manager/service/ManagerServiceImpl.java
@@ -87,7 +87,8 @@ public class ManagerServiceImpl implements ManagerService {
 			for (ServiceListItem serviceItem : req.getServiceTypes()) {
 				try {
 					ServiceType serviceTypeEnum = ServiceType.valueOf(serviceItem.getServiceType());
-					ManagerServiceType managerServiceType = ManagerServiceType.managerServiceType(manager, serviceTypeEnum);
+					ManagerServiceType managerServiceType = ManagerServiceType.managerServiceType(manager,
+						serviceTypeEnum);
 					managerServiceTypeRepository.save(managerServiceType);
 				} catch (IllegalArgumentException e) {
 					throw new BaseException(ResponseType.VALIDATION_FAILED);
@@ -218,7 +219,8 @@ public class ManagerServiceImpl implements ManagerService {
 			for (ServiceListItem serviceItem : req.getServiceTypes()) {
 				try {
 					ServiceType serviceTypeEnum = ServiceType.valueOf(serviceItem.getServiceType());
-					ManagerServiceType managerServiceType = ManagerServiceType.managerServiceType(manager, serviceTypeEnum);
+					ManagerServiceType managerServiceType = ManagerServiceType.managerServiceType(manager,
+						serviceTypeEnum);
 					managerServiceTypeRepository.save(managerServiceType);
 				} catch (IllegalArgumentException e) {
 					throw new BaseException(ResponseType.VALIDATION_FAILED);
@@ -253,48 +255,6 @@ public class ManagerServiceImpl implements ManagerService {
 		return ResponseDto.success();
 	}
 
-	@Override
-	public Page<ManagerListResponseDto> getManagerBypage(int page, int size) {
-		Pageable pageable = PageRequest.of(page, size);
-		return managerRepository.findAll(pageable)
-			.map(manager -> new ManagerListResponseDto(
-				manager.getName(),
-				manager.getUuid(),
-				manager.getId()
-			));
-	}
-
-	@Override
-	public ManagerResponseDto getManager(Long id){
-		Manager manager = managerRepository.findById(id).orElse(null);
-		System.out.println(manager.getId());
-		ManagerResponseDto managerResponseDto = ManagerResponseDto.builder()
-			.uuid(manager.getUuid())
-			.phoneNumber(manager.getPhoneNumber())
-			.name(manager.getName())
-			.birth(manager.getBirth())
-			.gender(manager.getGender())
-			.averageRate(manager.getAverageRate())
-			.region(manager.getRegions())
-			.isVerified(manager.getIsVerified())
-			.isDeleted(manager.getIsDeleted())
-			.build();
-		return managerResponseDto;
-	}
-	@Transactional
-	@Override
-	public void approveManager(Long managerId) {
-		Manager manager = managerRepository.findById(managerId).orElse(null);
-		manager.approve();
-	}
-
-	@Transactional
-	@Override
-	public void rejectManager(Long managerId) {
-		Manager manager = managerRepository.findById(managerId).orElse(null);
-		manager.reject();
-	}
-
 	// 리뷰 목록 조회
 	@Override
 	@Transactional(readOnly = true)
@@ -319,17 +279,6 @@ public class ManagerServiceImpl implements ManagerService {
 		);
 
 		return ResponseDto.success(responseDto);
-	}
-
-	@Override
-	public Page<ManagerListResponseDto> getManagerByPageWithStatus(int page, int size, Status status) {
-		Pageable pageable = PageRequest.of(page, size);
-		return managerRepository.findAllByIsVerified(status, pageable)
-			.map(manager -> new ManagerListResponseDto(
-				manager.getName(),
-				manager.getUuid(),
-				manager.getId()
-			));
 	}
 
 }
